added user_id property to the product model

+++ src/controller/productController.ts
@@ -4,7 +4,7 @@ import Product from "../models/productModel";
 
 //get all products
 const getProducts = asyncHandler(async (req: Request, res: Response) => {
-  const products = await Product.find();
+  const products = await Product.find({ user_id: req.user?.id });
   res.status(200).json(products);
 });
 
@@ -22,6 +22,8 @@ const getProduct = asyncHandler(async (req: Request, res: Response) => {
 
 // post/create a product
 const createProduct = asyncHandler(async (req: Request, res: Response) => {
+  // Log user object
+  console.log("User:", req.user);
   console.log("The Request body is:", req.body);
   const { name, description, price, category, manufacturer } = req.body;
   if (!name || !description || !price || !category || !manufacturer) {
@@ -34,6 +36,7 @@ const createProduct = asyncHandler(async (req: Request, res: Response) => {
     price,
     category,
     manufacturer,
+    user_id: req.user?.id,
   });
   res
     .status(201)
@@ -42,12 +45,19 @@ const createProduct = asyncHandler(async (req: Request, res: Response) => {
   //   res.status(201).json(product);
 });
 
-// update a product
+// update a particular product
 const updateProduct = asyncHandler(async (req, res) => {
   const product = await Product.findById(req.params.id);
   if (!product) {
     res.status(404);
     throw new Error("Product not found");
+  }
+
+  if (product.user_id.toString() !== req.user?.id) {
+    res.status(401);
+    throw new Error(
+      "Users don't have the permission to update other user's product"
+    );
   }
 
   const updatedProduct = await Product.findByIdAndUpdate(
@@ -56,15 +66,13 @@ const updateProduct = asyncHandler(async (req, res) => {
     { new: true }
   );
 
-  res
-    .status(200)
-    .json({
-      message: `Product updated for ${req.params.id}`,
-      product: updatedProduct,
-    });
+  res.status(200).json({
+    message: `Product updated for ${req.params.id}`,
+    product: updatedProduct,
+  });
 });
 
-// delete a product
+// delete a particular product
 const deleteProduct = asyncHandler(async (req: Request, res: Response) => {
   const product = await Product.findById(req.params.id);
   if (!product) {
@@ -72,14 +80,19 @@ const deleteProduct = asyncHandler(async (req: Request, res: Response) => {
     throw new Error("Product not found");
   }
 
+  if (product.user_id.toString() !== req.user?.id) {
+    res.status(401);
+    throw new Error(
+      "Users don't have the permission to update other user's product"
+    );
+  }
+
   await Product.deleteOne({ _id: product });
- 
-  res
-    .status(200)
-    .json({
-      message: `Product deleted for ${req.params.id}`,
-      product: product,
-    });
+
+  res.status(200).json({
+    message: `Product deleted for ${req.params.id}`,
+    product: product,
+  });
 });
 
 export { getProducts, getProduct, createProduct, updateProduct, deleteProduct };
